improve: exim: Address SAST (sonar) findings
- Remove static modifier on enum, which is implicitly static.
- Return logical comparisons directly instead of within another
conditional.
- Add private constructor to utility class (to prevent java from
implicitly offering a public constructor).
- Extract classes to reduce duplication.
